requirements.txt
flask
gunicorn
google-api-python-client
google-auth
google-auth-oauthlib
openai
psycopg2-binary
sqlalchemy
